# schemas/api/openapi.yaml
openapi: 3.1.0
info:
  title: ShakerProxy Control API
  version: 0.1.0-dev
  description: |
    Setup-safe API. Only operations implemented by this build are listed.

    Discover resources with GET /api/v1 and fetch this document from GET /api/v1/openapi.yaml.
    Every error, including unknown routes (404 not_found) and unsupported methods
    (405 method_not_allowed with an Allow header), uses the JSON shape
    {"error":{"code":"snake_case","message":"What went wrong and what to do next."}}.
    Request bodies are JSON (Content-Type application/json, optional charset=utf-8);
    decoding errors name the offending field.

    Authentication: sign in with POST /auth/login and send Authorization: Bearer <session_token>.
    Sessions slide (1 hour idle, 12 hours maximum); authenticated responses carry
    X-ShakerProxy-Session-Expires-At. Scoped API tokens (POST /auth/tokens) use the same header.

    Password confirmation: fields named "password" confirm the administrator. After any
    successful password check (including sign-in) the same session may omit "password" for
    10 minutes on routes marked "recent confirmation"; they answer 401 reauthentication_required
    when a password is needed. Routes marked "password every time" (network commit/confirm,
    deletion execution and retention, API token creation, enabling HTTPS decryption for every
    client, enabling decrypted HTTP body retention) answer 401 password_required without one.
servers:
  - url: /api/v1
paths:
...
  /openapi.yaml:
    get:
      operationId: getOpenAPIDocument
      description: This OpenAPI document as packaged with the appliance.
      responses:
        "200": {description: OpenAPI 3.1 document, content: {application/yaml: {schema: {type: string}}}}
        "503": {$ref: "#/components/responses/Error"}
  /auth/login:
    post:
      operationId: login
      description: Signs in the administrator. The session slides with use (1 hour idle, 12 hours maximum) and counts as a recent password confirmation for 10 minutes. Ten failed password attempts within five minutes answer 429 with Retry-After.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/LoginRequest"}}}}
      responses:
        "200": {description: Session token and expiry, content: {application/json: {schema: {$ref: "#/components/schemas/SessionResponse"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /auth/logout:
    post:
      operationId: logout
      description: Ends the session named by the bearer token. Idempotent; logged_out is false when the session was already gone.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Session revoked, content: {application/json: {schema: {type: object, additionalProperties: false, required: [schema, logged_out], properties: {schema: {const: 1}, logged_out: {type: boolean}}}}}}
        "401": {$ref: "#/components/responses/Error"}
  /auth/recover:
    post:
      operationId: recoverAdministrator
      description: Sets a new administrator password with one of the single-use recovery codes shown at setup (case, spaces and dashes are ignored). The new password must meet the setup password policy; a rejected password does not consume the code. All existing sessions are revoked and a new session is returned. Rate limited like sign-in.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/RecoverRequest"}}}}
      responses:
        "200": {description: New session and remaining code count, content: {application/json: {schema: {$ref: "#/components/schemas/RecoverResponse"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /auth/tokens:
    get:
      operationId: listAPITokens
      description: Session-only listing of hash-only, expiring API token metadata. Cleartext credentials and stored digests are never returned.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Token metadata and supported scopes, content: {application/json: {schema: {type: object, additionalProperties: false, required: [tokens, allowed_scopes, maximum_lifetime_seconds], properties: {tokens: {type: array, maxItems: 1024, items: {$ref: "#/components/schemas/APIToken"}}, allowed_scopes: {type: array, items: {$ref: "#/components/schemas/APITokenScope"}}, sensitive_scopes: {type: array, items: {$ref: "#/components/schemas/APITokenScope"}}, maximum_lifetime_seconds: {const: 7776000}}}}}}
        "401": {$ref: "#/components/responses/Error"}
    post:
      operationId: createAPIToken
      description: Session-only, freshly reauthenticated creation. The cleartext credential is returned exactly once and only its SHA-256 digest is persisted.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateAPITokenRequest"}}}}
      responses:
        "201": {description: Display-once credential and public metadata, headers: {X-ShakerProxy-Secret-Handling: {schema: {const: display-once}}}, content: {application/json: {schema: {$ref: "#/components/schemas/CreatedAPIToken"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
  /auth/tokens/{tokenID}:
    delete:
      operationId: revokeAPIToken
      description: Session-only, freshly reauthenticated and idempotent credential revocation.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/APITokenID"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/RevokeAPITokenRequest"}}}}
      responses:
        "200": {description: Revoked token metadata}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /integrations/forwarders:
    get:
      operationId: listForwarders
      description: Session-only status for bounded payload-free JSONL, HTTPS webhook, and syslog-over-TLS integrations. HMAC secrets are never returned after creation.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Forwarder status, queue lag/drop evidence, transport catalog, and immutable payload policy}
        "401": {$ref: "#/components/responses/Error"}
    post:
      operationId: createForwarder
      description: Creates a freshly reauthenticated integration in the disabled state. Webhook HMAC material is returned once. Local JSONL uses an appliance-owned bounded path; external transports reject non-public destinations and redirects.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateForwarderRequest"}}}}
      responses:
        "201": {description: Disabled integration and optional display-once HMAC secret, content: {application/json: {schema: {$ref: "#/components/schemas/CreatedForwarder"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
  /integrations/forwarders/{forwarderID}:
    patch:
      operationId: updateForwarder
      description: Recent confirmation. Changes the name, destination or event classes of a forwarder; kind and webhook secret are unchanged. expected_revision is optional (0 skips the check). A request matching the stored values returns the forwarder unchanged.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/ForwarderID"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdateForwarderRequest"}}}}
      responses:
        "200": {description: Updated public integration metadata, content: {application/json: {schema: {$ref: "#/components/schemas/ForwarderIntegration"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteForwarder
      description: Recent confirmation. Deletes a forwarder and its pending delivery queue. A local JSONL forwarder's output file is kept. The body is optional.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/ForwarderID"}]
      requestBody: {required: false, content: {application/json: {schema: {$ref: "#/components/schemas/DeleteForwarderRequest"}}}}
      responses:
        "200": {description: Deleted forwarder, content: {application/json: {schema: {type: object, additionalProperties: false, required: [schema, deleted, forwarder], properties: {schema: {const: 1}, deleted: {const: true}, forwarder: {$ref: "#/components/schemas/ForwarderIntegration"}}}}}}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /system/status:
    get:
      operationId: getSystemStatus
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Managed host state}
        "401": {$ref: "#/components/responses/Error"}
...
  /test-sessions:
    get:
      operationId: listTestSessions
      description: Newest-first test sessions. Requires an administrator session or devices:read.
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: device, schema: {type: string, maxLength: 128}, description: Any device reference.}
        - {in: query, name: state, schema: {type: string, enum: [RUNNING, STOPPED]}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 500, default: 50}}
      responses:
        "200": {description: Test sessions, content: {application/json: {schema: {$ref: "#/components/schemas/TestSessionList"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    post:
      operationId: startTestSession
      description: Starts a named test run against one device. The default name is "<device name> — YYYY-MM-DD HH:MM" (UTC). Starting a session stops the device's previous running session. capture true also starts a bounded header-only packet capture through the normal capture path; when capture is unavailable the session still starts and warnings explain why. Requires an administrator session or lab:write; no password prompt.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [device], properties: {device: {type: string, minLength: 1, maxLength: 128}, name: {type: string, maxLength: 120}, notes: {type: string, maxLength: 4096}, capture: {type: boolean, default: false}, full_capture: {type: boolean, default: false, description: "Record whole packets (implies capture) so TLS server names and certificates are analyzed."}}}}}}
      responses:
        "201": {description: Started test session, content: {application/json: {schema: {$ref: "#/components/schemas/TestSessionResult"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /test-sessions/{sessionID}:
    get:
      operationId: getTestSession
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/TestSessionID"}]
      responses:
        "200": {description: Test session, content: {application/json: {schema: {$ref: "#/components/schemas/TestSession"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
    patch:
      operationId: updateTestSession
      description: Renames a session or replaces its notes. Requires an administrator session or lab:write.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/TestSessionID"}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, minProperties: 1, properties: {name: {type: string, minLength: 1, maxLength: 120}, notes: {type: string, maxLength: 4096}}}}}}
      responses:
        "200": {description: Updated test session, content: {application/json: {schema: {$ref: "#/components/schemas/TestSession"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteTestSession
      description: Deletes session metadata only; captured traffic and packet files are kept. Deleting a running session stops its linked capture. Requires an administrator session or lab:write.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/TestSessionID"}]
      responses:
        "200": {description: Deleted, content: {application/json: {schema: {type: object, additionalProperties: false, required: [schema, id, deleted], properties: {schema: {const: 1}, id: {type: string}, deleted: {const: true}, warnings: {type: array, items: {type: string}}}}}}}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
...
  /traffic-policy:
    get:
      operationId: getTrafficPolicy
      description: Session-only current DNS and TLS traffic-policy document from the privileged host. The installed production gateway profile is required.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Current revisioned traffic policy, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyDocument"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    put:
      operationId: applyTrafficPolicy
      description: Freshly reauthenticated optimistic apply through the privileged gateway. Active controls require a confirmed IPv4 routed or single-arm plan, and listener failure leaves the packet path fail-open.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ApplyTrafficPolicyRequest"}}}}
      responses:
        "200": {description: Applied revisioned traffic policy, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyDocument"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /interception-ca/onboarding:
    get:
      operationId: getInterceptionCAOnboarding
      description: Where lab devices get the interception CA (http://<lab gateway>/, served only while HTTPS decryption is configured) and per-platform install steps and limitations. When the page is not reachable, available is false and reason says what to do next. Session or system:read token.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Onboarding status, content: {application/json: {schema: {$ref: "#/components/schemas/InterceptionCAOnboarding"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
...
  /capabilities:
    get:
      operationId: getCapabilities
      description: Returns the machine-validated feature registry, canonical glossary, and platform certification matrix. Capability claims are release metadata and never inferred from a running process alone.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Evidence-backed capability registry, content: {application/json: {schema: {$ref: "#/components/schemas/CapabilityBundle"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /cases:
    get:
      operationId: listCases
      description: Lists at most 1,024 durable cases newest-first. Case timelines are bounded and SHA-256 chained.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Durable case ledger, content: {application/json: {schema: {type: object, additionalProperties: false, required: [schema, cases], properties: {schema: {const: 1}, cases: {type: array, maxItems: 1024, items: {$ref: "#/components/schemas/Case"}}}}}}}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    post:
      operationId: createCase
      description: Creates an open case and its first custody-timeline entry. This does not create or apply an evidence hold.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateCaseRequest"}}}}
      responses:
        "201": {description: Created case, content: {application/json: {schema: {$ref: "#/components/schemas/Case"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /cases/{caseID}:
    get:
      operationId: getCase
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/CaseID"}]
      responses:
        "200": {description: Case detail and custody timeline, content: {application/json: {schema: {$ref: "#/components/schemas/Case"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    patch:
      operationId: updateCase
      description: Renames a case or edits its description (newlines allowed). expected_revision is optional (0 skips the check); reason defaults to a description of the change. A request matching the stored values returns the case without a new timeline entry.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/CaseID"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdateCaseRequest"}}}}
      responses:
        "200": {description: Updated case, content: {application/json: {schema: {$ref: "#/components/schemas/Case"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteCase
      description: Recent confirmation, session only. Deletes the case record and its timeline; evidence artifacts are kept. A case whose hold still protects a capture answers 409 case_hold_active; release the hold first. The body is optional.
      security: [{bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/CaseID"}]
      requestBody: {required: false, content: {application/json: {schema: {type: object, additionalProperties: false, properties: {expected_revision: {type: integer, minimum: 0}, password: {type: string, maxLength: 1024}}}}}}
      responses:
        "200": {description: Deleted case, content: {application/json: {schema: {type: object, additionalProperties: false, required: [schema, deleted, case], properties: {schema: {const: 1}, deleted: {const: true}, case: {$ref: "#/components/schemas/Case"}}}}}}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /preflight:
    get:
      operationId: inspectPreflight
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Read-only host and firewall-backend coexistence inspection}
        "401": {$ref: "#/components/responses/Error"}
...
  /captures:
    get:
      operationId: listCaptures
      description: Lists bounded capture sessions and current packet, drop, storage, closed-rotation analyzer-feed loss/error, and hash-manifest evidence.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Capture session list}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    post:
      operationId: startCapture
      description: Starts one bounded capture on the lab interface from the confirmed routed plan. Callers cannot select a host path, interface, service, command, or raw BPF expression.
      security: [{bearerAuth: []}]
      parameters:
        - {in: header, name: Idempotency-Key, required: true, schema: {type: string, minLength: 16, maxLength: 128, pattern: "^[A-Za-z0-9_-]+$"}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StartCaptureRequest"}
      responses:
        "201": {description: Capture worker accepted}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: Browser sessions authenticate every route. Expiring scoped API tokens authenticate only explicitly scoped read/automation routes and cannot manage credentials, forwarders, networking, deletion, exports, rules, or trust material.}
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, maxLength: 96}
        password: {type: string, maxLength: 1024}
...